pyproject.toml
[tool.poetry]
name = "mtl-escrow-python-sdk"
version = "0.0.1"
description = "A tiny Python SDK for Escrow.com API (unofficial)"
authors = ["MusicTech Lab <mariusz.smenzyk@musictechlab.io>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/musictechlab/mtl-escrow-python-sdk"
repository = "https://github.com/musictechlab/mtl-escrow-python-sdk"
keywords = ["escrow", "payments", "sdk", "api"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: Developers",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content :: CGI Tools/Libraries"
]
packages = [{include = "escrow_sdk"}]

[tool.poetry.dependencies]
python = ">=3.13,<4.0"
httpx = "^0.27.0"
pydantic = "^2.8.0"
typing-extensions = "^4.11.0"
flake8 = "^7.1.0"
pytest = "8.4.2"
coverage = "^7.10.7"
codecov = "^2.1.13"
black = "^25.9.0"
ruff = "^0.14.0"
autopep8 = "2.3.2"
click = "8.3.0"
python-dotenv = "^1.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"